Checks for game over, to force move to either win or block
